pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
where = ["src"]

[tool.hatch.version]
path = "src/ovos_solver_gemini_chat/__version__.py"

[project]
name = "ovos-solver-plugin-gemini-chat"
keywords = ["ovos", "skill", "plugin", "solver", "gemini", "chat"]
dynamic = ["version"]
authors = [
  { name="Flávio De Melo", email="flavio.eler@gmail.com" },
]
description = "A Gemini-based question solver plugin for OpenVoiceOS"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests<3.0.0,>=2.0.0",
    "google-genai>=1.38.0",
    "ovos-plugin-manager>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/femelo/ovos-solver-plugin-gemini-chat"
Issues = "https://github.com/femelo/ovos-solver-plugin-gemini-chat/issues"
